feat: new plugins DonationManager and SaDevWebsite
This is a two for one PR. Two new plugins!

SADevsWebsite: This plugin provides helper methods to manage the SA Devs
Website. Right now, it offers a context manager to get you a temporary
clone of the website's repo and a way to do a PR to the website.
Changing files is an exercise left up to the other plugin writer.

DonationManager is for SA Devs Season of Giving. It gives members a way
to record a donation with its ./donation command and gives admin tools
to manage the lifecycle of a submitted donation. Admins can approve,
change, and delete donations. Approved donations are pr'd to the
website in groups once every hour by a poller and a slack channel topic
is updated with our new donation total.